added memory-argument to run_conode (#841)
It provides an "-memory" option to the script so the conode server only has a fixed amount of memory to run. Once that amount is reached, it panics and the bash script will restart it after a timeout of 60sec. This is only a quick fix to the underlying problem of memory leakage that is going to be taken care of.