Make PubPoly serializable
share.PubPoly was not serializable, but we were writing it to disk
and it was coming back empty as a result. With this change, we
wrap it into a local structure which we know is serializable, and then
when it comes back from Storage, we call NewPubPoly on it to
get a share.PubPoly.

Closes #91