Make it possible to quickly test failure cases
Make BFTcosi timeouts adjustable. Make propagation timeouts
adjustable.

Adjust the timeouts to make the test finish fast.

Modify the skipchain failure test to use Pause/Unpause.
Write a 3rd block after Unpause. This is still flaky because
if the out of date server checks the signature, it refuses.
When that's fixed, this test needs to be strict instead
of allowing the failure as it does now.