Fixes so that suite can be changed at runtime
cothority.Suite defaults to P256 when vartime is enabled.

Fixes to make it work with P256 (i.e. places where the
length of a signature was hard coded).

Fixes to integration tests to follow these changes.
(But pop/test.sh and cisc/test.sh were already broken and
need to be fixed separately.)

Make run_conode.sh simpler, so that it can run from other
repositories as well unchanged.