Implement leader catchup
If a leader does not know of the latest block that a client
mentions, but it does know of the chain, it can ask others in the
chain's roster to catch it up.

Test it by nuking blocks from the db in order to simulate what would
happen if you started a leader with a database backup that was out of
date.

Also: Test follower catchup, which was already working correctly.