Finish removing SendRaw from skipchain
Change GetUpdateChain to operate on local blocks only. Delegate
responsibility to the client to go contact other servers according
to the latest roster (implemented in api.go, so that existing
clients do not see a change in behaviour).

Also: Added a test to check a worry I had about if new servers
which were invited into a chain would work correctly.

Fixes #1034.