Clean up status service, client
Clean up the type of the status reply, remove a call to an example
that will disappear with #1036.

Correctly display status replies which have multiple parts of the
server reporting stats.

Added status updater showing the size of the skipblock
database.

Also: fix various usability problems in scmgr uncovered while
testing this.

Fixes #940.