Container auto-setup and versioning, vgo experiment
If the conode_data directory does not have a private.toml file
then it will be created automatically.

Include the git revision hash in the Docker hub tag.

Add a rule for pushing it to Docker hub.

Build docker image with vgo. Have run_conode.sh use vgo
if it is available.