Responds empty commit when verification fail
Initially, the sub protocol would return an error immediately if the
verification function returns false and exit the protocol. As a result,
the sub-leader and the leader must wait for a timeout before they
proceed to the next step in the protocol. This commit brings an
exception handling mechanism inspired from bftcosi. The difference is
that instead of returning exceptions, the nodes return the commitment
without their own share and with their part of the mask set to false.