A hack to fix a panic we are seeing in the evoting demo
The underlying problem comes from issue #1244. The consequence of
that issue is that when a node that was down during some part of the
election fetches all ballots in order to try to verify the first shuffle
proposed by the leader, we end up with fewer ballots than the leader saw.
Having fewer ballots is no big deal, it just means we will refuse to sign
the forward link. Other conodes, which have the full chain, will sign it,
achieving a quorum.

If the election had only 2 ballots, we might trigger panics in the
shuffle library, and this change prevents that from happening.