Add contract for darc evolution
This commit adds a simple contract that checks the darc evolution
request. The functionality of Instruction.Action had to be changed to
allow this feature. That is, the darc-action for Invoke does not have
the "Invoke_" prefix anymore. This is unecessary because the re-created
darc.Request in an instruction must match the darc.Request during
verification. If the action in the instruction does match the action in
the darc, then verification will fail.

For example, when client submits an instruction, he/she actually signs
the `darc.Request` representation of the instruction. The request will
contain instruction.Action, which would be "Invoke__evolve", if the
client is invoking a contract with the "_evolve" command, in the old
implementation. This instruction is sent to the omniledger service and
it does two verifications, one in the service code and the other in the
contract.

(1) The service finds the corresponding darc for verification,
converts the instruction to a `darc.Request`, and calls
darc.Request.Verify. The verification will fail if the darc doesn't have
"Invoke__evolve" as one of its actions (a simple way to fix it is to
copy the expression from "_evolve" to "Invoke__evolve").

(2) We need to do the verification again in the contract because the
service verification doesn't try to decode the serialised darc. Hence,
the contract is given the same instruction, decodes it, and calls
darc.Verify. For this verification to work, the darc code will check
that the evolution is valid, e.g., the version is correct, base ID is
correct and most importantly, that the darc is signed by the owners of
the previous darc on the darc.Request representation, which *must* use
"_evolve" as its action.

Verification (1) works because we can add new rules to the darc. But
verification (2) will not work unless we make sure that the action in
darc.Request is re-created in the same way, from an instruction and from
a darc.

Another change is the Signer and Identity types. Before we always used
pointer types, which is unecessary because we never modify these structs
after creation. Now we use value types.