Choose the bucket id's better
This is an intermediate step on the way. Here we demonstrate
a new way of choosing id's for keys created by contracts.

Also, moved the hacky indexKey into "theEventLog", which
will soon be replaced with a proper spawn implementation
for contracts.

Bucket linking is currently disabled, but will be turned on next.