Add EventLogInstance class
The EventLogInstance is similar to the DarcInstance and ValueInstance
classes where it expects an OmniLedgerRPC object and uses that to
communicate with the OmniLedger service. We also remove the old EventLog
class because we no longer use a fully functional eventlog service,
everything RPC call (except Search) goes directly to OmniLedger.