Remove reference to broken and deprecated library
github.com/ericchiang/letsencrypt depends on
github.com/square/go-jose, which has recently changed to requiring an
import path ending in v2. ericchiang/letsencrypt now redirects
to ericchiang/go-acme, which was marked as deprecated on
12 Feb 2018.

So, cothority will not build because "go get -t ./..." returns
an error code when it fails to find go-jose.

Some possible fixes would be:
  1. change certificate.go to depend on golang.org/x/autocert/acme
  2. clone ericchiang/go-acme into dedis and fix it
  3. nuke letsencrypt support entirely because it is in CISC and
     CISC is not a priority for us now.

I have chosen fix 3 to fix the build right away, but we can consider
doing fix 1 if it looks to be easy.