Piggyback getTxs to sync omniledger state
We give nodes that were previously down the ability to catch-up to the
latest state. This is done by modifying the getTxs callback used in the
CollectTx protocol to take an additional argument - the latest
skipblock ID. When called, the node will check whether they are up to
date, if not, they'll attempt to synchronise with other nodes.

An earlier implementation maintained its own "background process" to
periodically try to catch-up. It has better encapsulation, but makes
extra communication and requires a lot more code change. Hence,
the piggybacking technique is used.