Implement view-change.v2
The new version of view-change replaces the previous version, which had
difficulty handling corner cases.

The design is similar to the view-change protocol in PBFT (OSDI99). We
keep the view-change message that replicas send when they detect an
anomaly. But we replace the new-view message with the ftcosi protocol
and block creation. The result of ftcosi is an aggregate signature of
all the nodes that agree to perform the view-change. The signature is
included in the block which nodes accept if the aggregate signature is
correct.

The technique above enables nodes to replay blocks to compute the most
up-to-date leader, which was not possible with the previous
implementation. Additionally, we are able to tolerate multiple failures,
which was also not possible previously. Finally, we enable view-change
by default and remove the possibility of switching it off because the
improvements above demonstrate its robustness.