Fix view-change bug and enable it
There is a bug that might cause new-view messages not being sent when an
anomaly is the last to arrive. This is fixed and is covered by
regression tests. Further, we fix some flakiness in the omniledger
service tests by waiting for skipblocks to finish propagating before
trying to close.