Make sure skipblock is consistent with collection
When we send a proof back to the client, the latest skipblock might not
be consistent with what is in the collection. That is, the block is
already stored in the skipchain database but updateCollectionCallback
hasn't been called yet. If this situation occurs, the proof will be
invalid. So add a check in GetProof to make sure this issue does not
happen.