Goodbye OmniLedger, hello ByzCoin
The name of the distributed ledger we are creating is now ByzCoin.
OmniLedger will be a system to coordinate cross-ledger transactions
allowing us to scale up the system by sharding the work across several
ByzCoin ledgers, so that transactions that are local to one ByzCoin
ledger do not need the services of OmniLedger to complete.