Proxy Authentication
Proxy Authentication is a new mechanism for Byzcoin
clients to get their transaction signed. Before, they had
to posess (and protect) a secret key. Now they can instead
be authenticated by a third-party identity provider (for
example OpenID Connect). The Authentication Proxy
takes the third part auth info and a message as input
verifies the authentication info, and then creates a
partial signature that binds the user id to the message.
The client then assembles these partial signatures into
a final signature that can be included with the transaction.

On the Darc verification side, the user-id binding
is checked with respect to the public key that corresponds
to all the private key shares in the auth proxies.