Forward-link signature roster
This prevents from using different roster than the same as the block
of the forward-link. It also prevents the verification to use a rotation
as the signature will use the correct roster and the mask will remove
failing nodes. In the case of a view change, this doesn't change anything
for the signature.

Fixes #1538