A toy to help understand base/height parameters
Use it like this: `BASE=3 HEIGHT=2 go test -run ShowBaseHeight`

By varying the BASE and HEIGHT env variables, you can see the
different ways that the forward links are created. You can
also see that len(fwd) == len(back), always.

Each run takes a few seconds because of all the collective
signing operations needed to make the skipchain. You can reduce
the number of blocks written from 200 to make it run faster, but
the number of blocks should be enough to cause all the forward
links to be created eventually.