Use new roster if aggregate is the same
This change introduces a way to use a newer roster that has the same
aggregate public key but with a different ID in startBFT. This is
because the newer roster is more likely to have nodes that are online at
the beginning of the list, so the subleaders in byzcoinx will be more
likely to be online.

This commit fixes the TestService_ViewChange2 failure on Jenkins.