Redefine the contract API to be an interface
The callback is now a factory function that turns the existing
instance data into a struct that implements byzcoin.Contract.

Contract writers are advised to have their contract structures embed
byzcoin.BasicContract, so that they only need to redefine
the methods they are interested in implementing.

Also: Removed InstructionsHash, which was derived data and not
useful to keep around.