Protobuf fixes: skipchain zigzag encoding and protoc version
Do not allow old versions of protoc, so that the .java files
do not get changed when we don't want them to be.

Fix the encoding of ints in skipchain.proto. If it was automatically
generated, proto.awk would have taken care of this.

Deploy a new version of cothority-js with the fixed protobuf
decoding rules.

Document the manually-maintained .proto files better.