conode: added call to rand.Seed in main()
We noticed that we are using rand.Perm in onet.Roster, but we never call
rand.Seed. In testing code this is perhaps helpful, but in production it's
not ideal. So we put the call to Seed in main(). Libtest.sh also uses
conode.go, so the integration tests are seeded too.

Fixes #1587.