Implement secure_darc contract
The old DARC contract has too much freedom. We need to limit this
freedom in two ways to improve security. First, the owner of any DARC
should not be allowed to change its rules arbitrarily, he/she is only
allowed to change the existing rules. Nevertheless, we still allow a set
of authorised parties to add rules. Second, we want to let certain DARCs
create new DARCs, but those new DARCs cannot have arbitrary rules.

These features are implemented in a new contract called "secure_darc".
For the first restriction, we changed the "evolve" command to reject
requests that tries to add new rules. We also added a new command
"evolve_unrestricted" to let authorised users add rules. For the second
restriction, we give an example of how to implement it statically in the
contract, which is to reject spawn requests that contains "spawn:darc".
Additionally, we explain in the example how to impelement it the same
feature dynamically.

Since we have more than one contract that can be decoded into a DARC, we
add a new parameter into ChainConfig to keep track of these contracts.

Fixes #1625