Verifier IDs in forward sig should not change
I tried looking into how to test this code but I didn't find a nice
way. We need to change StoreSkipBlock to test it. But I'm not aware of a
way to overload the skipchain service to change its StoreSkipBlock
function while keeping everything else working. The service name is used
in the service so just embedding skipchain.Service and then registering
it under a new name won't "just work". We should invest in a "cothority
testing framework" so that we can mock stuff easily.

Fixes #1243