Drop support for 32-bit binaries
In #1726 we looked into what it would take to make sure
that 32-bit servers worked, and it was too hard. So stop shipping
a binary that we know does not work anyway. During the evoting
project, we already discovered this, which is why the Windows binary
changed from Win32 to Win64 (see commits 770feadc and 6e4edb141b).