`bcadmin show` improvements
`bcadmin show` prints useful info to investigate the latest
block on different servers across the roster.

It no longer prints the genesis DARC from the bc config,
as that is not useful. If you want to see the genesis DARC
from the ledger, use `bcadmin darc show`.

Note: This is a backwards incompatible change to the behavior
of an admin tool.

Fixes #1755.