State change storage initialization and cleaning
This fixes some potential issues with the cleaning by size because
an array of sorted timestamp is used and its growth is not limited
and the maintainability depends on the insertion order. Cleaning is
now based on the block index.
It also removes the state changes reconstruction at conode boot by
a calculation of the current size as the state changes are already
built when a block is propagated and persisted in the db.