Change decodeKey into RecoverKey
A nice simplification is to move the DecodeKey function onto
the data type sent back from DecryptKey. That makes the order
of calls to the API clearer: after you get a DecryptKeyReply,
you can recover the key form it by adding the small amount of
additional information (your secret key).

Rewrote RecoverKey in order to not need the suite passed in.
Instead, it works with the points and scalars it has on hand
already.

Also remove an unused argument to AddRead.

Note: This is a backwards incompatible change, but this
API is not being actively used right now, and this is a dramatic
improvment in the API.