Rogue public-key attack defense
This PR adds a defense to the rogue public-key attack by adding
a check for the forward link level 0 so that when the roster is
changing with a new member, this member will need to provide a
proof of possession of the private key for each service registered
with a BN256 key pair. Each participant of the CoSi will need to
verify those proofs to accept the new block.

Fixes #1675