Fixes #1756
This PR implements a new function `StoreSkipBlockInternal` which is meant to be used in the case a service is used on the same host as the Skipchain service. I also removed a condition in `skipchain/skipchain.go` as discussed.

Note: I had to update 2 tests in `byzcoin/bcadmin/test.sh` line 95 and 98 that were supposed to pass. This is due to the removal of the condition. I am not 100% confident about the expected behavior now that I removed the condition,  so please double check.