bcadmin: adds --redirect for value contract
When using

```
$ bcadmin contract value spawn --value myValue
```

with the `--redirect` option, it does not spawn the value contract but redirects the transaction to stdout, later allowing this output to be used as the input of a spawn:deferred with pipe. (`bcadmin contract deferred spawn`)

This is a small step towards using bcadmin with deferred contract.