Change e-voting authentication model
Previously, e-voting authentication was designed to be
decentralised, with the signature being posted in the blocks
and the signature being checked during the verification of the
blocks by the conodes.

But because the signature was essentially a bearer token,
it must remain confidential, or else it can be used for
replay attacks.

With this change, the signature is now private data, which is
only known by the client and by the lead conode. The lead conode
becomes a centralised point of authentication, which is no worse
than our current system, which already has a centralised
authentication system (Tequila). In production, the same
organisation operates Tequila as the lead conode.

The process of shuffling and decrypting the votes is still
executed in a decentralised manner, with the private key
for an election still sharded across the cothority.