adds NoVerify option to GetProof
The current GetProof is very slow, as it verifies the whole proof at the end of the call.
This takes about 500ms on the test-system for 50 blocks, or 5s on the DEDIS-ByzCoin.

This PR adds a NoVerify field to the GetProof call, so that the verification is
skipped. Of course the verification still needs to be done at the client.