Require a signature from the lead conode
Make the fact that verifiers are counting on the lead conode to
check the user auth info more clear. Now for a given transaction
to verify, it must either be signed by the lead conode, or be one
of the 2 txn types where we expect the tx to be signed by one
of the other members of the cothority.