Fix signatures to be deterministic
Protobuf encoding of maps is not deterministic, as a result my previous change
was working for unit tests (which had no Name or Subtitle) but was not working
in real life. Fixed by making the message to sign in a deterministic manner.

Also address Gaylor's comment about locking, in the
worst possible way.