Improvements to E-voting (#1895)
* Use base verifier for evoting skipchains.

* Change e-voting authentication model

Previously, e-voting authentication was designed to be
decentralised, with the signature being posted in the blocks
and the signature being checked during the verification of the
blocks by the conodes.

But because the signature was essentially a bearer token,
it must remain confidential, or else it can be used for
replay attacks.

With this change, the signature is now private data, which is
only known by the client and by the lead conode. The lead conode
becomes a centralised point of authentication, which is no worse
than our current system, which already has a centralised
authentication system (Tequila). In production, the same
organisation operates Tequila as the lead conode.

The process of shuffling and decrypting the votes is still
executed in a decentralised manner, with the private key
for an election still sharded across the cothority.

* Require a signature from the lead conode

Make the fact that verifiers are counting on the lead conode to
check the user auth info more clear. Now for a given transaction
to verify, it must either be signed by the lead conode, or be one
of the 2 txn types where we expect the tx to be signed by one
of the other members of the cothority.

* Integrate the block number into the signature

In order to make the signature useless for replay,
we add the block number into the signed data.

* Fix signatures to be deterministic

Protobuf encoding of maps is not deterministic, as a result my previous change
was working for unit tests (which had no Name or Subtitle) but was not working
in real life. Fixed by making the message to sign in a deterministic manner.

Also address Gaylor's comment about locking, in the
worst possible way.

* Improve docs.