bcadmin: updates the link command
This update refactors the argument name and allows anyone to link to a roster given a darc (--darc) and an identity (--identity). This will create or update a config file that will be set as the BC envvar except if "--silent" is given.
It also add a new `bcadmin info` that prints information about the current BC config used.