Correct the ed25519_signer representation
Before: ed25519_signer hex values were represented as protobuf-encoded values
Now: ed25519_signer hex values are the raw MarshalBinary values

Why: this is how it is handled in the go-code. It also makes more sense, as the
hex-values are always preceded by 'ed25519:', so it is clear what kind of point
it is.