Parallel sending of messages
When using the cothority-ts in the omniledger-demonstrator I see sometimes a big
lag in the messages, specifically if dedis.nella.org is replying (I love that node
exactly for this reason!)

This PR solves the problem by sending all messages in parallel and returning the first
successful reply. It still tries to fetch all nodes, in case there are multiple
errors.