Use parallel
Uses the latest SendProtobufParallel from onet to simplify and speed up real-life usage of byzcoin. A lot of tests fail when sending commands to multiple nodes, mainly for racy-reasons. All failing tests got a cl.UseNode(0) to force the byzcoin.Client to send only to the leader.

See #2002 for a list of tests that fail and a first idea what might be wrong with the tests.