Implement StoreAllToReplica
Doing ForEach in the deferred contract might become expensive when the
state grows big. So we introduce StoreAllToReplica that makes a
"lightweight" copy of the trie and then applies the state changes. It
still makes copies of key/values but only those that are waiting to be
committed to a new block.