Atomically commit forward-link and new block
To avoid problems while storing the forward-link and the block, do it
atomically using StoreBlocks.

Also changes the tests with regard to waitPropagation. The
  byzcoin/bcadmin/lib/utils.go::WaitPropagation
method has been moved to
  byzcoin/api.go::Client.WaitPropagation
so that it can be used both by the lib and the tests.

Added second try in calypso test

This is a cherry-pick to v3 of #2122.