Refactors the processOneTx function
At first I only wanted to update an error in order to include the instanceID when one wants to create two times the same DARC. (the error I get is "contract darc tried to create existing instanceID", and I wanted to add the instanceID that already exist to this error).

I then refactored a bit the function that, IMHO, had some implementation details that made it harder to read (namely using a `defer` to add some post-instruction and the named return parameter, + some non-compliant formatting).