Various improvements to the BEvm client API code.
* Changed EvmContract to EvmContract and EvmContractInstance, the latter
being the result of a deployment on the former.

* NewEvmContract() now takes the ABI and bytecode as strings instead of
file names.

* Call() now returns a Go variable of the appropriate type (according to
the ABI), instead of requiring the caller to provide a good variable.
The caller still needs to know the type and cast it if needed, but that
makes his life easier as it can be e.g. checked by reflection.

* Adjusted the test code accordingly.