Separate personhood-contract and service
For the github.com/c4dt/byzcoin I want to include only the contracts from personhood,
but not the service. This PR puts the contracts in its own directory, so that the
byzcoin-repository can include only that.

Also rewrites the testing of the contracts to not use the service itself, but to
test only the contracts, without the whole byzcoin-service around. This could
also be applied to other contracts...