Optimize skipchain.GetUpdateChain
The javascript SkipchainRPC.getUpdateChain had a possible endless loop
when the fastest node returned a partially updated chain.
Also, when a node has been left out for some time, but then got re-integrated
in the chain, he might return a partially updated chain, which is not optimal.

This PR does the following:
- improves the skipchain-service so that the GetUpdateChain call returns the
longest chain a node can be sure of
- changes the skipchain-api-implementation to take correctly into account
a possible wrong node
- implements the javascript-skipchainRPC.getUpdateChain to correctly
update the chain in corner cases