Warn if wrong signature
With the change to the new signature scheme, it can happen that the
`ClienTransaction` is missing the version field, and the signing
will use the old version, and finally the verification will fail.

This adds a `log.Warn` if one of the signatures is wrong, and prints
a reference to `byzcoin.Client.CreateTransaction` in that case.