Support for editing elections before they start
The Open service will edit an election if the election
ID is set in the incoming Election structure and it
has not received any votes yet.

GetBox has been modified to make it possible to dump the
current state of an election.

Added support for this workflow in the evoting-admin tool.