Added support for translations for MoreInfo
It turns out that we need to send the voter to different URLs
depending on the language. This is implemented in a backwards
compatible way so that old elections can still be decoded.
Client code will need to implement backwards compatibility as
"if MoreInfoLang is empty, then use MoreInfo for all languages".