Use coinValue in spawner contract
Whenever a spawner contract creates a new coin instance, it fills it with all available coins.
This works fine if you only create one coin in a transaction. But if multiple instructions
with multiple spawners are created in the same transaction, it will fail.

This PR adds a new argument when spawning a coin that allows filling of the coin with only
a partial available amount.