Speed up proofs (#2228)
* Speed up proofs

In the latest dynacred2, there can be a lot of proofs requested when a new block
comes in. The previous code checked the forward-links for every proof, which
made it very slow.

This PR stores the latest trusted block, and doesn't check the forward-links
if a proof with that latest block is received.

Also changes new darcs with version=0 to have a random prevID, enabling
to store a darc with the same rules and description twice under a different
ID.

Also extends the Credential.setAttribute method to allow for strings in the value.