Add a caching system for the stateTries
The current stateTries have two bottlenecks:
- LoadConfigFromTrie - which is called everytime a LoadDarcFromTrie is called and
decodes the config protobuf structure every time anew
- LoadDarcFromTrie - which is called when evaluating signatures or authorizations

This PR adds a cache for both of these methods, so that calls within the
same block are cached. The test, which is a bit artificial, gives the
following improvements:
- LoadConfigFromTrie - 3x faster authorization check
- LoadDarcFromTrie - 4x faster authorization check

The latter is a bit artificial, because it measures looking up the same darcs
repeatedly, which in practice happens rarely.

The first test, however, is very relevant. In the 9cc3-chain, it is even worse,
as the config contains more nodes, which will make protobuf.Decode even
slower.