Patch up split and db
Two small patches for byzcoin:
- on 14th of May, another split occured, with higher-level forward-links pointing to
one block, and lower-level to another block. After restarting the nodes, it was still
in an unstable state, as the lower-level links led to another block than the higher-
level links. This patch makes sure that higher-level links that lead to nowhere are
removed from the db.
- part of the code spanned a lot of 'bbolt.Bucket's. This is fixed, and a method is
added to clean up