Rollup protocol
Previously the leader polled for new transactions. This is inefficient,
as new transactions have to wait for the leader to poll.

With this PR, the all followers forward transactions directly to the leader.
Once the leader has a transaction, he directly starts a new block.
Once the new block has been accepted, eventually queued up transactions
are also sent.

This has been tested to be able to go from old nodes to new nodes on a
local machine.