Re-arrange block updates and wait for new block (#2319)
The first call to ByzCoinRPC.instanceObservable was too late, making an eventual
recovery in ByzCoinRPC.getUpdate impossible.
ByzCoinRPC.getUpdate can fail if a block has been created but not yet sent to
the client. Now it waits for a confirmation of a new block and then retries
to get the updates.