Extract BCTest
This PR makes a nicer BCTest structure that can be used in the different tests.
Its main purpose was to have a correct 'CloseAll' method that waits for all
activity to stop.
But finally it also cleans up a lot of trailing cruft in different tests.

The goal is to have a BCTest structure that can also be used by other modules.
For this reason BCTest is exported in the main package.
For any idea how to make this cleaner, comments are welcome :)