Make getUpdates behave better
The previous ByzCoinRPC.getUpdate tried to be smart to cover an error in its use
in ByzCoinRPC.getUpdateInstances. The error in getUpdateInstances created a race
condition where two quickly updating blocks created two concurrent calls to
getUpdates which failed.

This PR adds a new field to the ByzCoin.GetUpdates RPC-call that allows it to
send updates with regard to a skipchain, and not with regards to a block.
This makes getUpdates much simpler.
Furthermore it removes the race in getUpdateInstances and waits for each
update before accepting new blocks.