Using dynacred logins for bcadmin
The user management with dynacred is much easier and allows using of a web-frontend
to have multiple devices and recovery by a group you trust.
This PR links the dynacred login with the basic functions given by bcadmin,
so that you can use the private key given by a 'Add Device' in the
web-frontend.