Checking roster for new nodes
The byzcoin contract did not correctly check for new nodes in a new proposed
roster, this led to a wrong node being stored in the chain.
The bcadmin did not correctly load the public files of a new node, a
node to be deleted, or a node for a leader change.

Both are fixed in this PR.

It depends on dedis/onet#676